Adding int4range type, and enabling Test-Kitchen-driven development
Adding the int4range type.

Also, given that epgsql requires a running postgres server with a very
specific configuration, and some specific test data pre-loaded, I've
integrated it with TK to make it easy to work on it.

Now running the test suite is as delightful as running `kitchen verify`!

Future work: put it on travis. Might do when I find some spare time.